backend/foodgram/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.13
requests==2.31.0
urllib3==2.0.4
yarg==0.1.9
Django==3.2.19
django-filter==22.1
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
python-dotenv==1.0.0
gunicorn==20.1.0
flake8==6.0.0
flake8-isort==6.0.0
django-cors-headers==3.13.0
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.14.0
djoser==2.1.0
django-filters==2.4.0
reportlab
sqlparse==0.4.4